Added getOne method and refactored countries method in CountryResource class
