+++ src/main/java/com/github/elcioishizuka/citiesapi/countries/repository/CountryRepositoty.java
@@ -1,7 +1,8 @@
-package com.github.elcioishizuka.citiesapi.repository;
+package com.github.elcioishizuka.citiesapi.countries.repository;
 
 import com.github.elcioishizuka.citiesapi.countries.Country;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CountryRepositoty extends JpaRepository<Country, Long> {
+
 }
