+++ src/main/resources/application.properties
@@ -1,3 +1,8 @@
 spring.datasource.url=jdbc:postgresql://localhost:5432/cities
 spring.datasource.username=postgres_user_city
 spring.datasource.password=super_password
+
+# Show at log window the sql commands used to execute the requested by browser or Postman
+spring.jpa.show-sql=true
+
+spring.jpa.properties.hibernate.format_sql=true
